+++ src/main/java/com/shipper/service/EquipmentService.java
@@ -32,32 +32,57 @@ public class EquipmentService {
 	}
 
 	public void deleteEquipment(Long id) {
-		equipmentRepo.deleteById(id);
+		if (MessageValidator.isPositiveNumber(id)) {
+			equipmentRepo.deleteById(id);
+		}
+		else {
+			System.err.print(id + " is not a positive number");
+		}
 	}
 
 	public Long createEquipment(String equipmentName, int quantity, Long id) {
-		deliveryList = deliveryRepository.findFirstById(id);
+		if (MessageValidator.isValidEquipment(equipmentName, quantity, id)) {
+			deliveryList = deliveryRepository.findFirstById(id);
 		Equipment eq = new Equipment(equipmentName,quantity,deliveryList);
 		equipmentRepo.save(eq);
 		return eq.getId();
+		} else {
+			System.err.print("Message is not valid!");
+			return (long) 0;
+		}
+		
 	}
 	public List<Equipment> getEquipments(Long id) {
+		if (MessageValidator.isPositiveNumber(id)) {
 		return  equipmentRepo.findAllByDeliveryListId(id);
+		}else {
+			System.err.print(id + " is not a positive number");
+			return equipmentRepo.findAll();
+		}
 	}
 	
 	public void createDeliveryList(String name, String eq) {
-		DeliveryList delivery = new DeliveryList(name,"Halux");
-		deliveryRepository.save(delivery);
-		
+		if (MessageValidator.areStringsValid(name, eq)) {
+				DeliveryList delivery = new DeliveryList(name,"Halux");
+		deliveryRepository.save(delivery);	
 		Equipment equipment = new Equipment(eq, 6,delivery);
 		equipmentRepo.save(equipment);
+		} else {
+			System.err.print("Message is not valid!");
+		}
+	
 	}
 	
 	public void updateList(Message message) {
-		equipment = equipmentRepo.findFirstById(message.getId());
+		if (MessageValidator.isValidEquipment(message.getName(),message.getQuantity(), message.getId())) {
+				equipment = equipmentRepo.findFirstById(message.getId());
 		equipment.setName(message.getName());
 		equipment.setQuantity(message.getQuantity());
 		equipmentRepo.save(equipment);
+		} else {
+			System.err.print("Message is not valid!");
+		}
+	
 	}
 	
 //	@PostConstruct
